simplify get_fields, remove redundant code
